+++ CHANGELOG.md
@@ -5,7 +5,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.17.2-rc6]
 ### Changed
+ - We now avoid scaling utilization and clamp it directly to the appropriate
+   range.
+
+## [0.17.2-rc5]
+### Changed
+ - We now clamp CPU utilization to number of cores, not scaled to.
+
+## [0.17.2-rc4]
+### Changed
+ - We now elide sampling threads that are not also processes in the observer.
+
+## [0.17.2-rc3]
+### Changed
+ - Correct `rss` to be reported in terms of bytes, not pages, as before.
+
+## [0.17.2-rc2]
+### Removed
+- Observer no longer emits tick data for kernel and user-space time.
+### Changed
+- Observer now distinguishes between parent and children processes.
 - Config can now be specified using an env var `LADING_CONFIG`. If set, the env var takes precedence over the on-disk config file.
 
 ## [0.17.2-rc1]